tpdi/tp/requirements.txt
accelerate
streamlit 
flask
opencv-python
numpy
pandas
scikit-learn 
Pillow==10.3.0
transformers>=4.38.0
tokenizers==0.15.2
diffusers==0.25.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
moviepy==2.1.2
numba==0.60.0
huggingface_hub==0.21.0
numba
rich


